Chat creation logic changed + websocket
New chats are no longer created seperately from messages meaning a chat will not be created until a message is sent within the chat itself. This new creation method has is implemented in post as well as websocket. When a new chat is created by a user and then they send a message in that chat the chat will then be created along with the message on the receivers side.